Add files via upload
"# cyclistic_analysis" This project was developed as a capstone for the Google Analytics Professional Certification program. It entails a detailed analysis of rideshare data over 12 months from the fictional company Cyclistic. The goal of this analysis was to define important differences between the two types of riders casual and member - in order to help guide the company in future ad campaigns with their stated objective of converting casual riders into members.

*Data can be found at https://divvy-tripdata.s3.amazonaws.com/index.html